Add files via upload
Leverage guest data and booking behaviour patterns to devise a strategy for hotel revenue management
using Data science and Machine learning.s. This Hotel Booking Cancellation model can be useful not only for the vacationer but for 
the hotel’s owners. The data was divided into two parts i.e. the training and the test dataset. Models 
were developed based on the training dataset and applied to the test dataset to find out the accuracy of 
each model based on the predicted values generated